feat(webapp): send mail notififcation of sale transactions
