+++ packages/webapp/package.json
@@ -20,6 +20,13 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
     "@testing-library/user-event": "^7.2.1",
+    "@tiptap/extension-color": "latest",
+    "@tiptap/extension-text-style": "2.1.13",
+    "@tiptap/core": "2.1.13",
+    "@tiptap/pm": "2.1.13",
+    "@tiptap/extension-list-item": "2.1.13",
+    "@tiptap/react": "2.1.13",
+    "@tiptap/starter-kit": "2.1.13",
     "@types/jest": "^26.0.15",
     "@types/js-money": "^0.6.1",
     "@types/lodash": "^4.14.172",
